First checkpoint in making cheesy-arena support a more generic game, with two gamepieces (each scored in their own structure), making it easier to customize and use the FMS for Team 766's Mechanical M-Ayhem game.
remove support for realtime scoring via the PLC (to avoid updating it for each year's game; will not be used for M-Ayhem).
remove support for team signs in field - not used for M-Ayhem.
remove support for reading/writing data from The Blue Alliance and for Nexus - not used for M-Ayhem.

These changes should make it easier to customize the FMS for each year's M-Ayhem game, while ideally also making the effort to keep this fork up to date with the upstream Cheesy Arena code reasonable.